requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
python-multipart
pydantic
bcrypt
passlib
python-jose
email-validator
psycopg2
uvicorn
fastapi-pagination